adding some routes to post controller and adding additional routes to the next app

+++ app/src/main/resources/db/migration/V2__create_post_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE post (
-    id INT PRIMARY KEY AUTO_INCREMENT,
+    id BIGINT PRIMARY KEY AUTO_INCREMENT,
     user_id BIGINT,
     title VARCHAR(255),
     content TEXT,
